Move DataModules from torchgeo.trainers to torchgeo.datasets (#220)
