+++ experiments/test_chesapeakecvpr_models.py
@@ -11,7 +11,8 @@ import os
 import pytorch_lightning as pl
 import torch
 
-from torchgeo.trainers import ChesapeakeCVPRDataModule, ChesapeakeCVPRSegmentationTask
+from torchgeo.datasets import ChesapeakeCVPRDataModule
+from torchgeo.trainers import ChesapeakeCVPRSegmentationTask
 
 ALL_TEST_SPLITS = [["de-val"], ["pa-test"], ["ny-test"], ["pa-test", "ny-test"]]
