+++ setup.cfg
@@ -38,6 +38,8 @@ install_requires =
     pillow>=2.9
     # pyproj 2.2+ required for CRS object
     pyproj>=2.2
+    # pytorch-lightning 1.3+ required for gradient_clip_algorithm argument to Trainer
+    pytorch-lightning>=1.3
     # rasterio 1.0.16+ required for CRS support
     rasterio>=1.0.16
     # rtree 0.5+ required for 3D index support
@@ -71,8 +73,6 @@ datasets =
 train =
     # omegaconf 2.1+ required for to_object method
     omegaconf>=2.1
-    # pytorch-lightning 1.3+ required for gradient_clip_algorithm argument to Trainer
-    pytorch-lightning>=1.3
     # scikit-learn 0.18+ required for sklearn.model_selection module
     scikit-learn>=0.18
     # segmentation-models-pytorch 0.2+ required for smp.losses module
