+++ tests/datasets/test_bigearthnet.py
@@ -13,7 +13,7 @@ from _pytest.fixtures import SubRequest
 from _pytest.monkeypatch import MonkeyPatch
 
 import torchgeo.datasets.utils
-from torchgeo.datasets import BigEarthNet
+from torchgeo.datasets import BigEarthNet, BigEarthNetDataModule
 
 
 def download_url(url: str, root: str, *args: str, **kwargs: str) -> None:
@@ -107,3 +107,33 @@ class TestBigEarthNet:
         "to automaticaly download the dataset."
         with pytest.raises(RuntimeError, match=err):
             BigEarthNet(str(tmp_path))
+
+
+class TestBigEarthNetDataModule:
+    @pytest.fixture(scope="class", params=zip(["s1", "s2", "all"], [True, True, False]))
+    def datamodule(self, request: SubRequest) -> BigEarthNetDataModule:
+        bands, unsupervised_mode = request.param
+        root = os.path.join("tests", "data", "bigearthnet")
+        batch_size = 1
+        num_workers = 0
+        dm = BigEarthNetDataModule(
+            root,
+            bands,
+            batch_size,
+            num_workers,
+            unsupervised_mode,
+            val_split_pct=0.3,
+            test_split_pct=0.3,
+        )
+        dm.prepare_data()
+        dm.setup()
+        return dm
+
+    def test_train_dataloader(self, datamodule: BigEarthNetDataModule) -> None:
+        next(iter(datamodule.train_dataloader()))
+
+    def test_val_dataloader(self, datamodule: BigEarthNetDataModule) -> None:
+        next(iter(datamodule.val_dataloader()))
+
+    def test_test_dataloader(self, datamodule: BigEarthNetDataModule) -> None:
+        next(iter(datamodule.test_dataloader()))
