+++ tests/datasets/test_cyclone.py
@@ -14,7 +14,7 @@ from _pytest.fixtures import SubRequest
 from _pytest.monkeypatch import MonkeyPatch
 from torch.utils.data import ConcatDataset
 
-from torchgeo.datasets import TropicalCycloneWindEstimation
+from torchgeo.datasets import CycloneDataModule, TropicalCycloneWindEstimation
 
 
 class Dataset:
@@ -91,3 +91,25 @@ class TestTropicalCycloneWindEstimation:
     def test_not_downloaded(self, tmp_path: Path) -> None:
         with pytest.raises(RuntimeError, match="Dataset not found or corrupted."):
             TropicalCycloneWindEstimation(str(tmp_path))
+
+
+class TestCycloneDataModule:
+    @pytest.fixture(scope="class")
+    def datamodule(self) -> CycloneDataModule:
+        root = os.path.join("tests", "data", "cyclone")
+        seed = 0
+        batch_size = 1
+        num_workers = 0
+        dm = CycloneDataModule(root, seed, batch_size, num_workers)
+        dm.prepare_data()
+        dm.setup()
+        return dm
+
+    def test_train_dataloader(self, datamodule: CycloneDataModule) -> None:
+        next(iter(datamodule.train_dataloader()))
+
+    def test_val_dataloader(self, datamodule: CycloneDataModule) -> None:
+        next(iter(datamodule.val_dataloader()))
+
+    def test_test_dataloader(self, datamodule: CycloneDataModule) -> None:
+        next(iter(datamodule.test_dataloader()))
