+++ tests/datasets/test_naip.py
@@ -12,7 +12,7 @@ import torch.nn as nn
 from _pytest.monkeypatch import MonkeyPatch
 from rasterio.crs import CRS
 
-from torchgeo.datasets import NAIP, BoundingBox, ZipDataset
+from torchgeo.datasets import NAIP, BoundingBox, NAIPChesapeakeDataModule, ZipDataset
 
 
 class TestNAIP:
@@ -50,3 +50,27 @@ class TestNAIP:
             IndexError, match="query: .* not found in index with bounds:"
         ):
             dataset[query]
+
+
+class TestNAIPChesapeakeDataModule:
+    @pytest.fixture(scope="class")
+    def datamodule(self) -> NAIPChesapeakeDataModule:
+        dm = NAIPChesapeakeDataModule(
+            os.path.join("tests", "data", "naip"),
+            os.path.join("tests", "data", "chesapeake", "BAYWIDE"),
+            batch_size=2,
+            num_workers=0,
+        )
+        dm.patch_size = 32
+        dm.prepare_data()
+        dm.setup()
+        return dm
+
+    def test_train_dataloader(self, datamodule: NAIPChesapeakeDataModule) -> None:
+        next(iter(datamodule.train_dataloader()))
+
+    def test_val_dataloader(self, datamodule: NAIPChesapeakeDataModule) -> None:
+        next(iter(datamodule.val_dataloader()))
+
+    def test_test_dataloader(self, datamodule: NAIPChesapeakeDataModule) -> None:
+        next(iter(datamodule.test_dataloader()))
