+++ tests/datasets/test_sen12ms.py
@@ -12,7 +12,7 @@ from _pytest.fixtures import SubRequest
 from _pytest.monkeypatch import MonkeyPatch
 from torch.utils.data import ConcatDataset
 
-from torchgeo.datasets import SEN12MS
+from torchgeo.datasets import SEN12MS, SEN12MSDataModule
 
 
 class TestSEN12MS:
@@ -82,3 +82,26 @@ class TestSEN12MS:
             ds = SEN12MS(root, bands=bands, checksum=False)
             x = ds[0]["image"]
             assert x.shape[0] == len(bands)
+
+
+class TestSEN12MSDataModule:
+    @pytest.fixture(scope="class", params=["all", "s1", "s2-all", "s2-reduced"])
+    def datamodule(self, request: SubRequest) -> SEN12MSDataModule:
+        root = os.path.join("tests", "data", "sen12ms")
+        seed = 0
+        bands = request.param
+        batch_size = 1
+        num_workers = 0
+        dm = SEN12MSDataModule(root, seed, bands, batch_size, num_workers)
+        dm.prepare_data()
+        dm.setup()
+        return dm
+
+    def test_train_dataloader(self, datamodule: SEN12MSDataModule) -> None:
+        next(iter(datamodule.train_dataloader()))
+
+    def test_val_dataloader(self, datamodule: SEN12MSDataModule) -> None:
+        next(iter(datamodule.val_dataloader()))
+
+    def test_test_dataloader(self, datamodule: SEN12MSDataModule) -> None:
+        next(iter(datamodule.test_dataloader()))
