--- tests/trainers/test_resisc45.py
@@ -1,30 +1,0 @@
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License.
-
-import os
-
-import pytest
-
-from torchgeo.trainers import RESISC45DataModule
-
-
-@pytest.fixture(scope="module")
-def datamodule() -> RESISC45DataModule:
-    root = os.path.join("tests", "data", "resisc45")
-    batch_size = 2
-    num_workers = 0
-    dm = RESISC45DataModule(root, batch_size, num_workers)
-    dm.prepare_data()
-    dm.setup()
-    return dm
-
-
-class TestRESISC45DataModule:
-    def test_train_dataloader(self, datamodule: RESISC45DataModule) -> None:
-        next(iter(datamodule.train_dataloader()))
-
-    def test_val_dataloader(self, datamodule: RESISC45DataModule) -> None:
-        next(iter(datamodule.val_dataloader()))
-
-    def test_test_dataloader(self, datamodule: RESISC45DataModule) -> None:
-        next(iter(datamodule.test_dataloader()))
