+++ torchgeo/datasets/__init__.py
@@ -5,7 +5,7 @@
 
 from .advance import ADVANCE
 from .benin_cashews import BeninSmallHolderCashews
-from .bigearthnet import BigEarthNet
+from .bigearthnet import BigEarthNet, BigEarthNetDataModule
 from .cbf import CanadianBuildingFootprints
 from .cdl import CDL
 from .chesapeake import (
@@ -13,6 +13,7 @@ from .chesapeake import (
     Chesapeake7,
     Chesapeake13,
     ChesapeakeCVPR,
+    ChesapeakeCVPRDataModule,
     ChesapeakeDC,
     ChesapeakeDE,
     ChesapeakeMD,
@@ -21,9 +22,9 @@ from .chesapeake import (
     ChesapeakeVA,
     ChesapeakeWV,
 )
-from .cowc import COWC, COWCCounting, COWCDetection
+from .cowc import COWC, COWCCounting, COWCCountingDataModule, COWCDetection
 from .cv4a_kenya_crop_type import CV4AKenyaCropType
-from .cyclone import TropicalCycloneWindEstimation
+from .cyclone import CycloneDataModule, TropicalCycloneWindEstimation
 from .etci2021 import ETCI2021
 from .eurosat import EuroSAT
 from .geo import (
@@ -35,7 +36,7 @@ from .geo import (
     ZipDataset,
 )
 from .gid15 import GID15
-from .landcoverai import LandCoverAI
+from .landcoverai import LandCoverAI, LandcoverAIDataModule
 from .landsat import (
     Landsat,
     Landsat1,
@@ -50,15 +51,15 @@ from .landsat import (
     Landsat9,
 )
 from .levircd import LEVIRCDPlus
-from .naip import NAIP
+from .naip import NAIP, NAIPChesapeakeDataModule
 from .nwpu import VHR10
 from .patternnet import PatternNet
-from .resisc45 import RESISC45
-from .sen12ms import SEN12MS
+from .resisc45 import RESISC45, RESISC45DataModule
+from .sen12ms import SEN12MS, SEN12MSDataModule
 from .sentinel import Sentinel, Sentinel2
-from .so2sat import So2Sat
+from .so2sat import So2Sat, So2SatDataModule
 from .spacenet import SpaceNet, SpaceNet1, SpaceNet2, SpaceNet4
-from .ucmerced import UCMerced
+from .ucmerced import UCMerced, UCMercedDataModule
 from .utils import BoundingBox, collate_dict
 from .zuericrop import ZueriCrop
 
@@ -77,6 +78,7 @@ __all__ = (
     "ChesapeakeVA",
     "ChesapeakeWV",
     "ChesapeakeCVPR",
+    "ChesapeakeCVPRDataModule",
     "Landsat",
     "Landsat1",
     "Landsat2",
@@ -89,31 +91,40 @@ __all__ = (
     "Landsat8",
     "Landsat9",
     "NAIP",
+    "NAIPChesapeakeDataModule",
     "Sentinel",
     "Sentinel2",
     # VisionDataset
     "ADVANCE",
     "BeninSmallHolderCashews",
     "BigEarthNet",
+    "BigEarthNetDataModule",
     "COWC",
     "COWCCounting",
     "COWCDetection",
+    "COWCCountingDataModule",
     "CV4AKenyaCropType",
     "ETCI2021",
     "EuroSAT",
     "GID15",
     "LandCoverAI",
+    "LandcoverAIDataModule",
     "LEVIRCDPlus",
     "PatternNet",
     "RESISC45",
+    "RESISC45DataModule",
     "SEN12MS",
+    "SEN12MSDataModule",
     "So2Sat",
+    "So2SatDataModule",
     "SpaceNet",
     "SpaceNet1",
     "SpaceNet2",
     "SpaceNet4",
     "TropicalCycloneWindEstimation",
+    "CycloneDataModule",
     "UCMerced",
+    "UCMercedDataModule",
     "VHR10",
     "ZueriCrop",
     # Base classes
