+++ torchgeo/trainers/ucmerced.py
@@ -3,137 +3,10 @@
 
 """UC Merced trainer."""
 
-from typing import Any, Dict, Optional
-
-import pytorch_lightning as pl
-import torch
-import torchvision.transforms.functional
-from torch.nn.modules import Conv2d, Linear
-from torch.utils.data import DataLoader
-from torchvision.transforms import Compose, Normalize
-
-from ..datasets import UCMerced
 from .tasks import ClassificationTask
-
-# https://github.com/pytorch/pytorch/issues/60979
-# https://github.com/pytorch/pytorch/pull/61045
-DataLoader.__module__ = "torch.utils.data"
-Conv2d.__module__ = "nn.Conv2d"
-Linear.__module__ = "nn.Linear"
 
 
 class UCMercedClassificationTask(ClassificationTask):
     """LightningModule for training models on the UC Merced Dataset."""
 
     num_classes = 21
-
-
-class UCMercedDataModule(pl.LightningDataModule):
-    """LightningDataModule implementation for the UC Merced dataset.
-
-    Uses random train/val/test splits.
-    """
-
-    band_means = torch.tensor([0, 0, 0])  # type: ignore[attr-defined]
-
-    band_stds = torch.tensor([1, 1, 1])  # type: ignore[attr-defined]
-
-    def __init__(
-        self,
-        root_dir: str,
-        batch_size: int = 64,
-        num_workers: int = 4,
-        **kwargs: Any,
-    ) -> None:
-        """Initialize a LightningDataModule for UCMerced based DataLoaders.
-
-        Args:
-            root_dir: The ``root`` arugment to pass to the UCMerced Dataset classes
-            batch_size: The batch size to use in all created DataLoaders
-            num_workers: The number of workers to use in all created DataLoaders
-        """
-        super().__init__()  # type: ignore[no-untyped-call]
-        self.root_dir = root_dir
-        self.batch_size = batch_size
-        self.num_workers = num_workers
-
-        self.norm = Normalize(self.band_means, self.band_stds)
-
-    def preprocess(self, sample: Dict[str, Any]) -> Dict[str, Any]:
-        """Transform a single sample from the Dataset.
-
-        Args:
-            sample: dictionary containing image
-
-        Returns:
-            preprocessed sample
-        """
-        sample["image"] = sample["image"].float()
-        sample["image"] /= 255.0
-        c, h, w = sample["image"].shape
-        if h != 256 or w != 256:
-            sample["image"] = torchvision.transforms.functional.resize(
-                sample["image"], size=(256, 256)
-            )
-        sample["image"] = self.norm(sample["image"])
-        return sample
-
-    def prepare_data(self) -> None:
-        """Make sure that the dataset is downloaded.
-
-        This method is only called once per run.
-        """
-        UCMerced(self.root_dir, download=True, checksum=False)
-
-    def setup(self, stage: Optional[str] = None) -> None:
-        """Initialize the main ``Dataset`` objects.
-
-        This method is called once per GPU per run.
-
-        Args:
-            stage: stage to set up
-        """
-        transforms = Compose([self.preprocess])
-
-        self.train_dataset = UCMerced(self.root_dir, "train", transforms=transforms)
-        self.val_dataset = UCMerced(self.root_dir, "val", transforms=transforms)
-        self.test_dataset = UCMerced(self.root_dir, "test", transforms=transforms)
-
-    def train_dataloader(self) -> DataLoader[Any]:
-        """Return a DataLoader for training.
-
-        Returns:
-            training data loader
-        """
-        return DataLoader(
-            self.train_dataset,
-            batch_size=self.batch_size,
-            num_workers=self.num_workers,
-            shuffle=True,
-        )
-
-    def val_dataloader(self) -> DataLoader[Any]:
-        """Return a DataLoader for validation.
-
-        Returns:
-            validation data loader
-        """
-        return DataLoader(
-            self.val_dataset,
-            batch_size=self.batch_size,
-            num_workers=self.num_workers,
-            shuffle=False,
-        )
-
-    def test_dataloader(self) -> DataLoader[Any]:
-        """Return a DataLoader for testing.
-
-        Returns:
-            testing data loader
-        """
-        return DataLoader(
-            self.test_dataset,
-            batch_size=self.batch_size,
-            num_workers=self.num_workers,
-            shuffle=False,
-        )
